+++ src/hooks/index.ts
@@ -2,6 +2,7 @@ export { useAquariumData } from './useAquariumData'
 export { useCreateAquarium } from './useCreateAquarium';
 export { useLoadAquariumDetails } from './useLoadAquariumDetails';
 export { useLoadAquariums } from './useLoadAquariums';
+export { useLogIn } from './useLogIn';
 export { useResetParams } from './useResetParams';
 export { useSelectedAquarium } from './useSelectedAquarium';
 export { useSignUp } from './useSignUp';
